Fix sign extension bug on x,y,z variables

+++ HMC5883L/HMC5883L.pde
@@ -86,7 +86,7 @@ void setup()
  */
 void loop()
 {
-    int x,y,z; //triple axis data
+    int16_t x,y,z; //triple axis data
 
     //Tell the HMC5883 where to begin reading data
     Wire.beginTransmission(HMC_ADDRESS);
